Split config to core config parts and project parts Add dynamic section modules import from config directory path specified in pytest.ini Make config_parser package public
