+++ common/facade/libs/__init__.py
@@ -1,2 +1,2 @@
 from common._libs.helpers.singleton import *
-from common._libs.config_parser.config_dto import APIValidationDTO
+from common.config_parser.config_dto import APIValidationDTO
