+++ pytest.ini
@@ -1,2 +1,2 @@
 [pytest]
-config_dir = C:\Users\test\Documents\PyCats\config
+config_dir = ./config
